Refactor monetary values to use int64 (cents) across the application
- Updated product use case tests to convert prices to cents using money.ToCents.
- Modified shipping use case to handle monetary values in cents.
- Adjusted discount entity to store min order value and max discount value as int64 (cents).
- Changed order entity to store total amount, shipping cost, discount amount, and final amount as int64 (cents).
- Updated payment transaction entity to use int64 for transaction amounts.
- Refactored shipping rate entity to store base rate, min order value, and free shipping threshold as int64 (cents).
- Enhanced money utility functions to apply percentages to cents values.
- Adjusted payment services and repositories to handle int64 amounts.
- Updated API handlers to convert monetary values to and from cents.
- Modified mock repositories to reflect changes in monetary value handling.

+++ internal/domain/repository/payment_transaction_repository.go
@@ -31,5 +31,5 @@ type PaymentTransactionRepository interface {
 	CountSuccessfulByOrderIDAndType(orderID uint, transactionType entity.TransactionType) (int, error)
 
 	// SumAmountByOrderIDAndType sums the amount of transactions of a specific type for an order
-	SumAmountByOrderIDAndType(orderID uint, transactionType entity.TransactionType) (float64, error)
+	SumAmountByOrderIDAndType(orderID uint, transactionType entity.TransactionType) (int64, error)
 }
